+++ src/test/php/AbstractMethodTest.php
@@ -94,7 +94,6 @@ class AbstractMethodTest extends TestCase
 
     /**
      * @test
-     * @doesNotPerformAssertions
      */
     public function amountOfCallsToMethodIsZeroIfNotCalled(): void
     {
@@ -103,7 +102,6 @@ class AbstractMethodTest extends TestCase
 
     /**
      * @test
-     * @doesNotPerformAssertions
      */
     public function recordsAmountOfCallsToMethod(): void
     {
@@ -114,7 +112,6 @@ class AbstractMethodTest extends TestCase
 
     /**
      * @test
-     * @doesNotPerformAssertions
      */
     public function optionalArgumentsCanNotBeVerifiedWhenNotExplicitlyPassed(): void
     {